refactor: relocate Outcome type-aliases
