+++ src/main/kotlin/nz/adjmunro/knomadic/KnomadicAliases.kt
@@ -21,33 +21,6 @@ import nz.adjmunro.knomadic.raise.RaiseScope
 public typealias KotlinResult<T> = Result<T>
 
 /**
- * Alias for a [RaiseScope] context runner used for [Outcome].
- *
- * *If injecting lambdas into [outcomeOf], consider using this as the lambda
- * type to inherit the [RaiseScope] context for your own lambdas.*
- */
-@KnomadicDsl
-public typealias OutcomeScope<Ok, Error> = RaiseScope<Error>.() -> Ok
-
-/**
- * Alias for a [RaiseScope] context runner used for [Maybe].
- *
- * *If injecting lambdas into [maybeOf], consider using this as the lambda
- * type to inherit the [RaiseScope] context for your own lambdas.*
- */
-@KnomadicDsl
-public typealias MaybeScope<Ok> = RaiseScope<Any>.() -> Ok
-
-/**
- * Alias for a [RaiseScope] context runner used for [Faulty].
- *
- * *If injecting lambdas into [faultyOf], consider using this as the lambda
- * type to inherit the [RaiseScope] context for your own lambdas.*
- */
-@KnomadicDsl
-public typealias FaultyScope<Error> = RaiseScope<Error>.() -> Unit
-
-/**
  * Alias for a [Flow] of [Fetch] statuses.
  * @see fetch
  */
@@ -128,50 +101,6 @@ public typealias OutcomeFetch<Ok, Error> = Flow<Fetch<Outcome<Ok, Error>>>
 @KnomadicDsl
 public typealias ResultFetch<Ok> = Flow<Fetch<KotlinResult<Ok>>>
 
-/**
- * Alias for a [flow][Flow] of a [faulty][Faulty] result.
- *
- * ```kotlin
- * val faulty: FaultyFlow<String> = flow {
- *     // Emit a Faulty from inside the Flow
- *     emit(faultyOf { "An error occurred." })
- * }
- * ```
- *
- * @see Faulty
- */
-@KnomadicDsl
-public typealias FaultyFlow<Error> = Flow<Faulty<Error>>
-
-/**
- * Alias for a [flow][Flow] of a [maybe][Maybe] result.
- *
- * ```kotlin
- * val maybe: MaybeFlow<String> = flow {
- *     // Emit a Maybe from inside the Flow
- *     emit(maybeOf { "A value." })
- * }
- * ```
- *
- * @see Maybe
- */
-@KnomadicDsl
-public typealias MaybeFlow<Ok> = Flow<Maybe<Ok>>
-
-/**
- * Alias for a [flow][Flow] of an [outcome][Outcome] result.
- *
- * ```kotlin
- * val outcome: OutcomeFlow<String, Int> = flow {
- *     // Emit an Outcome from inside the Flow
- *     emit(outcomeOf { 42 })
- * }
- * ```
- *
- * @see Outcome
- */
-@KnomadicDsl
-public typealias OutcomeFlow<Ok, Error> = Flow<Outcome<Ok, Error>>
 
 /**
  * Alias for a [flow][Flow] of a [result][KotlinResult].
