+++ src/main/kotlin/nz/adjmunro/knomadic/outcome/flow/FoldOutcomeFlow.kt
@@ -2,7 +2,7 @@ package nz.adjmunro.knomadic.outcome.flow
 
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
-import nz.adjmunro.knomadic.OutcomeFlow
+import nz.adjmunro.knomadic.outcome.OutcomeFlow
 import nz.adjmunro.knomadic.outcome.Outcome
 import nz.adjmunro.knomadic.outcome.OutcomeDsl
 import nz.adjmunro.knomadic.outcome.members.collapse
